requirements.txt
optuna
pandas
scikit-learn
torch_geometric
numpy
matplotlib
networkx
scipy